# faster iteration over a `Flatten` of heterogenous iterators