@nanosoldier 2025-06-18
[Your benchmark job](https://github.com/JuliaLang/julia/pull/58522#issuecomment-2981437329) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/master/benchmark/by_hash/9293799_vs_aa10603/report.md).