@adienes 2025-07-02
anybody have a good idea for a test here? just looking at the `code_lowered` I see about twice as many appearances of `dynamic` which I assume is relevant but I don't know really what the best way to test for that is.